extracted the expressions into a new file
